refactor(signatures): improve AsfaloadSignatureTrait error mapping of minisign's errors
Following this comment [1], we improve the mapping of minisign's PError
to our different cases. It appears that currently the code was not
necessary (we sem to only generate IO PErrors), but it is more explicit
and a good learning.

Note that it required to switch to a version of minisign that is not yet
published as a crate. We now use minisign from the github repo as a
specific commit.

1: https://github.com/asfaload/asfasign/pull/7#discussion_r2303046705