refactor(signer_file_types): move KeyFormat to signatures crate
I want to add key_format function to the AsfaloadPublicKeyTrait, to
return the format of the key used. To do this, the enum has to be
defined in the signatures crate, but it would be a duplicate of the
signers_file_types definition, so I moved it.
The downside is that I had to make the signatures crate use serde.
Another particularity is that signers_file_types republishes the
KeyFormat type with `pub use signatures::keys::KeyFormat`. I did this so
I didn't have to edit all imports of KeyFormat from signers_file_types.
We'll see later if this is a bad idea.....