feat(common): define SignedFileWithKind
This was introduced together with a SignedFile<KindMarker>.
A SignedFile was already defined without a marker, but it appeared it
would be useful for the user lib. So SignedFile's definition was moved
to common, adding a marker type in the process.
Some methods were added in the process to make it usable from current
code (eg to access to the wrapped instance's field).