feat(signers_file): add timestamp field and remove initial_version
Having a timestamp will allow additional consistency checks and avoid
replay attacks (eg copying an old signers config as a new one on the
server).

initial_version was a left-over from earlier explorations, but is not
working: how can the url of the file'initial version field be in it if
the url include the hash of the commit introduing that file?