feat(signatures): multi algorithms with non-generic types
This is a big refactoring. It was due to a problem that appeared in the
previous commit in the auth crate. Introducing a dummy algorithm
implementing the Asfaload*Traits, it appeared the code was not as
generic as should be: lots of places now required to specify the generic
type arguments. This tries to fix that problem.

It introduces enums for public keys, secret keys, signatures and
keypairs. Each enum has one case per signing algorithm supported. This
is still only one, but introducing these enums allows to hide the
generic type arguments to the code using it.

These enum simply wrap the corresponding Asfaload type (eg
AsfaloadPublicKey<_>), setting its generic type, and implement the same
traits, delegating the action of the trait's functions to the wrapped
value.
For example, AsfaloadPublicKeys has one case Minisign which wrap
AsfaloadPublicKey<minisign::PublicKey>. It implements
AsfaloadPublicKeyTrait.